+++ Program.cs
@@ -4,8 +4,10 @@ using System.Text.Json.Serialization;
 using Qotion;
 using Qotion.Client;
 using Qotion.OneBot;
-using Qotion.OneBot.API;
-using Qotion.OneBot.Event;
+using OntBotAPI = Qotion.OneBot.API;
+using OneBotEvent =  Qotion.OneBot.Event;
+using OneBotRequest = Qotion.OneBot.Request;
+using OneBotResponse = Qotion.OneBot.Response;
 
 Console.WriteLine("Please enter your account:");
 var account = Int64.Parse(Console.ReadLine() ?? string.Empty);
@@ -27,13 +29,12 @@ void ReceivedMessage(object sender, string data)
     if(postType!="message") return;
     var messageType = json["message_type"].ToString();
     if(messageType!="private") return;
-    var privateMessage = Event.Convert<PrivateMessageEvent>(data);
-    var message = new SendMessage("private", account, 0, $"你发送的信息是{privateMessage.message}，现在是{DateTime.Now}。信息自动处理完成。", true);
-    var request = new Request<SendMessage>("send_private_msg",message);
+    var privateMessage = OneBotResponse.Response.Convert<OneBotEvent.PrivateMessage>(data);
+    var message = new OntBotAPI.SendMessage("private", account, 0, $"你发送的信息是{privateMessage.message}，现在是{DateTime.Now}。信息自动处理完成。", true);
+    var request = new OneBotRequest.Request<OntBotAPI.SendMessage>("send_private_msg",message);
     wsApi.OnMessage += (o, s) => Console.WriteLine(s);
     wsApi.Open();
-    var requestJson = JsonSerializer.Serialize(request);
-    requestJson = requestJson.Replace("_params", "params");
+    var requestJson = OneBotRequest.Request.Serialize(request);
     Console.WriteLine("send:"+requestJson);
     wsApi.Send(requestJson);
 }